requirement.txt
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.1.0
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
cryptography==46.0.4
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.3
fqdn==1.5.1
fsspec==2026.1.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
ipykernel==7.1.0
ipython==8.38.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
lark==1.3.1
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.2.0
more-itertools==10.8.0
mpmath==1.3.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.5.3
notebook_shim==0.2.4
numba==0.63.1
numpy==2.2.6
ollama==0.6.1
openai-whisper==20250625
overrides==7.7.0
packaging==26.0
pandocfilters==1.5.1
parso==0.8.5
pillow==12.1.0
platformdirs==4.5.1
prometheus_client==0.24.1
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.2
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
rsa==4.9.1
Send2Trash==2.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
tiktoken==0.12.0
tinycss2==1.4.0
tomli==2.4.0
torch==2.10.0
tornado==6.5.4
tqdm==4.67.2
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.3
wcwidth==0.5.3
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
whisper==1.1.10
